Rename XTypeVisitor to TypeVisitor, change package
